Move ProjectSpec into MengeCore
The ability to parse a Menge project no longer lies in the main app, but
in the core library. This facilitates parsing projects into arbitrary
contexts.

TODO:
1) Update VS2015 builds
2) Update linux builds
