+++ projects/VS2013/Menge/MengeCore.vcxproj
@@ -361,6 +361,7 @@ xcopy /y $(TargetPath) $(ExeDir)
     <ClCompile Include="$(SrcDir)\MengeCore\Agents\Events\EventTriggerFactory.cpp" />
     <ClCompile Include="$(SrcDir)\MengeCore\Agents\Events\EventEffectAgentState.cpp" />
     <ClCompile Include="$(SrcDir)\MengeCore\Agents\Events\TargetAgentById.cpp" />
+    <ClCompile Include="$(SrcDir)\MengeCore\ProjectSpec.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="$(SrcDir)\mengeCore\Core.h" />
@@ -559,6 +560,7 @@ xcopy /y $(TargetPath) $(ExeDir)
     <ClInclude Include="$(SrcDir)\MengeCore\Agents\Events\EventTriggerExternal.h" />
     <ClInclude Include="$(SrcDir)\MengeCore\Agents\Events\EventEffectAgentState.h" />
     <ClInclude Include="$(SrcDir)\MengeCore\Agents\Events\TargetAgentById.h" />
+    <ClInclude Include="$(SrcDir)\MengeCore\ProjectSpec.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="tinyxml_lib.vcxproj">
