+++ projects/VS2013/Menge/MengeCore.vcxproj.filters
@@ -621,6 +621,9 @@
     <ClCompile Include="$(SrcDir)\MengeCore\Agents\Events\TargetAgentById.cpp">
       <Filter>Source Files\Agents\Events</Filter>
     </ClCompile>
+    <ClCompile Include="$(SrcDir)\MengeCore\ProjectSpec.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="$(SrcDir)\mengeCore\Core.h">
@@ -1210,6 +1213,9 @@
     </ClInclude>
     <ClInclude Include="$(SrcDir)\MengeCore\Agents\Events\TargetAgentById.h">
       <Filter>Header Files\Agents\Events</Filter>
+    </ClInclude>
+    <ClInclude Include="$(SrcDir)\MengeCore\ProjectSpec.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
